create precision subpackage
