+++ pytorch_lightning/accelerators/plugins/precision/__init__.py
@@ -1,0 +1,4 @@
+from pytorch_lightning.accelerators.plugins.precision.apex_amp import ApexMixedPrecisionPlugin
+from pytorch_lightning.accelerators.plugins.precision.mixed import MixedPrecisionPlugin
+from pytorch_lightning.accelerators.plugins.precision.native_amp import NativeMixedPrecisionPlugin
+from pytorch_lightning.accelerators.plugins.precision.precision_plugin import PrecisionPlugin
